requirements.txt
# Production dependencies
fastapi
uvicorn
python-dotenv
pydantic
pydantic-settings

# Development dependencies
black
flake8
httpx
pytest
pytest-asyncio
